LayoutTests/fast/dom/StyleSheet/stylesheet-symbol-names.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("This tests document.styelSheets with Symbol property names.");

styleElement1 = document.createElement("style");
styleElement1.setAttribute("id", "");
document.head.appendChild(styleElement1);

styleElement2 = document.createElement("style");
styleElement2.setAttribute("id", "test");
document.head.appendChild(styleElement2);

shouldBe("document.styleSheets.length", "3");
shouldBe("document.styleSheets['']", "undefined");
shouldBe("document.styleSheets['test']", "styleElement2.sheet");
shouldBe("document.styleSheets[Symbol()]", "undefined");
shouldNotThrow("delete document.styleSheets['test']");
shouldNotThrow("delete document.styleSheets[Symbol()]");
shouldBe("document.styleSheets.length", "3");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
