LayoutTests/fast/dom/StyleSheet/stylesheet-symbol-names-expected.txt

This tests document.styelSheets with Symbol property names.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.styleSheets.length is 3
PASS document.styleSheets[''] is undefined
PASS document.styleSheets['test'] is styleElement2.sheet
PASS document.styleSheets[Symbol()] is undefined
PASS delete document.styleSheets['test'] did not throw exception.
PASS delete document.styleSheets[Symbol()] did not throw exception.
PASS document.styleSheets.length is 3
PASS successfullyParsed is true

TEST COMPLETE
